collector-python/requirements.txt
# collector-python/requirements.txt
pika
requests
python-dotenv